+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## NEXT
+* Fixed PKCS10 Certification Request Info's attribute canonicalisation
 * Allow `Asn1` builder unary `+` for transparent wrappers around `Asn1Element`/`Asn1Encodable`, and for serializable
   values when a `Der` instance is in context.
 * Change the `X509AlgorithmIdentifier` constructor to take a single nullable `parameters` element.